+++ codegen/smithy-dafny-codegen/src/main/java/software/amazon/polymorph/smithyrust/generator/RustLibraryShimGenerator.java
@@ -10,10 +10,14 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
 import software.amazon.polymorph.traits.DafnyUtf8BytesTrait;
 import software.amazon.polymorph.traits.LocalServiceTrait;
+import software.amazon.polymorph.traits.PositionalTrait;
 import software.amazon.polymorph.traits.ReferenceTrait;
 import software.amazon.polymorph.utils.IOUtils;
 import software.amazon.polymorph.utils.MapUtils;
@@ -624,6 +628,21 @@ public class RustLibraryShimGenerator extends AbstractRustShimGenerator {
       .getBindingShape(operationShape)
       .get();
     if (bindingShape.isServiceShape()) {
+      StructureShape inputShape = operationIndex.getInputShape(operationShape).get();
+      final String rustValue;
+      final boolean isRustOption;
+      if (inputShape.hasTrait(PositionalTrait.class)) {
+        MemberShape onlyMember = PositionalTrait.onlyMember(inputShape);
+        rustValue = "input." + toSnakeCase(onlyMember.getMemberName());
+        isRustOption = true;
+      } else {
+        rustValue = "input";
+        isRustOption = false;
+      }
+      variables.put("inputToDafny", toDafny(inputShape, rustValue, isRustOption, false).toString());
+
+      StructureShape outputShape = operationIndex.getOutputShape(operationShape).get();
+      variables.put("outputFromDafny", fromDafny(outputShape, "inner_result.value()", false, false).toString());
       variables.put(
         "operationSendBody",
         IOUtils.evalTemplate(
@@ -1126,24 +1145,40 @@ public class RustLibraryShimGenerator extends AbstractRustShimGenerator {
     final String operationModuleName = toSnakeCase(
       operationName(operationShape)
     );
-    final String outerContent = IOUtils.evalTemplate(
-      getClass(),
-      "runtimes/rust/conversions/operation.rs",
-      variables
-    );
+
+    Optional<StructureShape> inputStructure = operationIndex.getOutputShape(operationShape);
+    final boolean hasInputStructure = inputStructure.isPresent() && !inputStructure.get().hasTrait(PositionalTrait.class);
+    Optional<StructureShape> outputStructure = operationIndex.getOutputShape(operationShape);
+    final boolean hasOutputStructure = outputStructure.isPresent() && !outputStructure.get().hasTrait(PositionalTrait.class);
+
+    final Set<String> childModules = new HashSet<>();
+    if (hasInputStructure) {
+      childModules.add("_" + variables.get("snakeCaseSyntheticOperationInputName"));
+    }
+    if (hasOutputStructure) {
+      childModules.add("_" + variables.get("snakeCaseSyntheticOperationOutputName"));
+    }
     final RustFile outerModule = new RustFile(
       Path.of("src", "conversions", operationModuleName + ".rs"),
-      TokenTree.of(outerContent)
+      declarePubModules(childModules.stream())
     );
 
-    final RustFile requestModule = operationRequestConversionModule(
-      operationShape
-    );
-    final RustFile responseModule = operationResponseConversionModule(
-      operationShape
-    );
+    Set<RustFile> result = new HashSet<>(Set.of(outerModule));
 
-    return Set.of(outerModule, requestModule, responseModule);
+    if (hasInputStructure) {
+      final RustFile requestModule = operationRequestConversionModule(
+        operationShape
+      );
+      result.add(requestModule);
+    }
+    if (hasOutputStructure) {
+      final RustFile responseModule = operationResponseConversionModule(
+        operationShape
+      );
+      result.add(responseModule);
+    }
+
+    return result;
   }
 
   @Override
@@ -1231,6 +1266,7 @@ public class RustLibraryShimGenerator extends AbstractRustShimGenerator {
       structureId,
       StructureShape.class
     );
+    final boolean isPositional = structureShape.hasTrait(PositionalTrait.class);
     final Map<String, String> variables = MapUtils.merge(
       serviceVariables(),
       operationVariables(operationShape),
@@ -1241,25 +1277,47 @@ public class RustLibraryShimGenerator extends AbstractRustShimGenerator {
       fluentMemberSettersForStructure(structureShape).toString()
     );
 
-    // unwrap() is safe as long as the builder is infallible
-    return TokenTree.of(
-      evalTemplate(
-        """
-        #[allow(dead_code)]
-        pub fn from_dafny(
-            dafny_value: ::std::rc::Rc<
-                crate::r#$dafnyTypesModuleName:L::$structureName:L,
-            >,
-        ) -> crate::operation::$snakeCaseOperationName:L::$rustStructureName:L {
-            crate::operation::$snakeCaseOperationName:L::$rustStructureName:L::builder()
-                $fluentMemberSetters:L
-                .build()
-                .unwrap()
-        }
-        """,
-        variables
-      )
-    );
+    if (isPositional) {
+      return TokenTree.of(
+        evalTemplate(
+          """
+            #[allow(dead_code)]
+            pub fn from_dafny(
+                dafny_value: ::std::rc::Rc<
+                    crate::r#$dafnyTypesModuleName:L::$structureName:L,
+                >,
+            ) -> crate::operation::$snakeCaseOperationName:L::$rustStructureName:L {
+                crate::operation::$snakeCaseOperationName:L::$rustStructureName:L::builder()
+                    $fluentMemberSetters:L
+                    .build()
+                    .unwrap()
+            }
+            """,
+          variables
+        )
+      );
+    } else {
+
+      // unwrap() is safe as long as the builder is infallible
+      return TokenTree.of(
+        evalTemplate(
+          """
+            #[allow(dead_code)]
+            pub fn from_dafny(
+                dafny_value: ::std::rc::Rc<
+                    crate::r#$dafnyTypesModuleName:L::$structureName:L,
+                >,
+            ) -> crate::operation::$snakeCaseOperationName:L::$rustStructureName:L {
+                crate::operation::$snakeCaseOperationName:L::$rustStructureName:L::builder()
+                    $fluentMemberSetters:L
+                    .build()
+                    .unwrap()
+            }
+            """,
+          variables
+        )
+      );
+    }
   }
 
   private RustFile wrappedModule() {
@@ -1393,136 +1451,6 @@ public class RustLibraryShimGenerator extends AbstractRustShimGenerator {
       "%s::%s".formatted(qualifiedRustServiceErrorType(), rustErrorName)
     );
     return variables;
-  }
-
-  protected String unionName(final UnionShape unionShape) {
-    return unionShape.getId().getName(service);
-  }
-
-  protected String rustUnionName(final UnionShape unionShape) {
-    return toPascalCase(unionName(unionShape));
-  }
-
-  protected String qualifiedRustUnionName(final UnionShape unionShape) {
-    return "%s::types::%s".formatted(
-        topLevelNameForShape(unionShape),
-        rustUnionName(unionShape)
-      );
-  }
-
-  /**
-   * Generates values for variables commonly used in union-specific templates.
-   */
-  protected HashMap<String, String> unionVariables(
-    final UnionShape unionShape
-  ) {
-    final HashMap<String, String> variables = new HashMap<>();
-    final String unionName = unionName(unionShape);
-    variables.put("unionName", unionName);
-    variables.put("snakeCaseUnionName", toSnakeCase(unionName));
-    variables.put("dafnyUnionName", unionName);
-    variables.put("rustUnionName", rustUnionName(unionShape));
-    variables.put("qualifiedRustUnionName", qualifiedRustUnionName(unionShape));
-    return variables;
-  }
-
-  protected String unionMemberName(final MemberShape memberShape) {
-    return memberShape.getMemberName();
-  }
-
-  protected String rustUnionMemberName(final MemberShape memberShape) {
-    return toPascalCase(unionMemberName(memberShape));
-  }
-
-  protected String dafnyUnionMemberName(final MemberShape memberShape) {
-    return unionMemberName(memberShape);
-  }
-
-  /**
-   * Generates values for variables commonly used in union-member-specific templates.
-   */
-  protected HashMap<String, String> unionMemberVariables(
-    final MemberShape memberShape
-  ) {
-    final String memberName = unionMemberName(memberShape);
-    final Shape targetShape = model.expectShape(memberShape.getTarget());
-
-    final HashMap<String, String> variables = new HashMap<>();
-    variables.put("unionMemberName", memberName);
-    variables.put("snakeCaseUnionMemberName", toSnakeCase(memberName));
-    variables.put("dafnyUnionMemberName", dafnyUnionMemberName(memberShape));
-    variables.put("rustUnionMemberName", rustUnionMemberName(memberShape));
-    variables.put("unionMemberType", rustTypeForShape(targetShape));
-    return variables;
-  }
-
-  private String rustTypeForShape(final Shape originalShape) {
-    // First handle indirection like @reference
-    final ModelUtils.ResolvedShapeId resolvedShapeId = ModelUtils.resolveShape(
-      originalShape,
-      model
-    );
-    final Shape shape = model.expectShape(resolvedShapeId.resolvedId());
-
-    return switch (shape.getType()) {
-      case BOOLEAN -> "::std::primitive::bool";
-      // integral
-      case BYTE -> "::std::primitive::i8";
-      case SHORT -> "::std::primitive::i16";
-      case INTEGER -> "::std::primitive::i32";
-      case LONG -> "::std::primitive::i64";
-      // floats
-      case FLOAT -> "::std::primitive::f32";
-      case DOUBLE -> "::std::primitive::f64";
-      // special numerics
-      case BIG_INTEGER -> "::num::bigint::BigInt";
-      case BIG_DECIMAL -> "::num::rational::BigRational";
-      // special collections
-      case BLOB -> "::aws_smithy_types::Blob";
-      case STRING -> {
-        //noinspection deprecation
-        if (shape.hasTrait(EnumTrait.class)) {
-          yield qualifiedRustEnumType(
-            ModelUtils.stringToEnumShape(shape.asStringShape().orElseThrow())
-          );
-        }
-        yield "::std::string::String";
-      }
-      case ENUM -> qualifiedRustEnumType(shape.asEnumShape().orElseThrow());
-      // other simple shapes
-      case TIMESTAMP -> "::aws_smithy_types::DateTime";
-      // aggregates
-      case STRUCTURE -> {
-        final StructureShape structureShape = (StructureShape) shape;
-        yield qualifiedRustStructureType(structureShape);
-      }
-      case LIST -> {
-        final ListShape listShape = (ListShape) shape;
-        final String memberType = rustTypeForShape(
-          model.expectShape(listShape.getMember().getTarget())
-        );
-        yield "::std::vec::Vec<%s>".formatted(memberType);
-      }
-      case MAP -> {
-        final MapShape mapShape = (MapShape) shape;
-        final String keyType = rustTypeForShape(
-          model.expectShape(mapShape.getKey().getTarget())
-        );
-        final String valueType = rustTypeForShape(
-          model.expectShape(mapShape.getValue().getTarget())
-        );
-        yield "::std::collections::HashMap<%s, %s>".formatted(
-            keyType,
-            valueType
-          );
-      }
-      case UNION -> qualifiedRustUnionName((UnionShape) shape);
-      case RESOURCE -> qualifiedRustResourceType((ResourceShape) shape);
-      case SERVICE -> qualifiedRustServiceType((ServiceShape) shape);
-      default -> throw new UnsupportedOperationException(
-        "Unsupported shape type: " + shape.getType()
-      );
-    };
   }
 
   protected boolean isRustFieldRequired(
