+++ codegen/smithy-dafny-codegen/src/main/resources/templates/runtimes/rust/operation/outer_send_body.rs
@@ -1,4 +1,4 @@
-        let inner_input = crate::conversions::$snakeCaseOperationName:L::_$snakeCaseSyntheticOperationInputName:L::to_dafny(input);
+        let inner_input = $inputToDafny:L;
         let inner_result =
             ::dafny_runtime::md!(client.dafny_client.clone()).$operationName:L(&inner_input);
         if matches!(
@@ -6,9 +6,7 @@
             crate::r#_Wrappers_Compile::Result::Success { .. }
         ) {
             Ok(
-                crate::conversions::$snakeCaseOperationName:L::_$snakeCaseSyntheticOperationOutputName:L::from_dafny(
-                    inner_result.value().clone(),
-                ),
+                $outputFromDafny:L,
             )
         } else {
             Err(crate::conversions::error::from_dafny(
